Recover a crashed run's task-folder changes at the next launch
The pre-run snapshot lived only in the worker's memory, so a run that
ASTRA did not live to finish was marked interrupted at the next launch
with nothing to compare, and every file it touched fell out of its
record.

The worker now keeps the baseline in the task folder's `diagnostics/`
(`task_folder_baseline_<run id>.json`) for the length of the run, and
`recordChanges` removes it. Content fingerprints are SHA-256 prefixes
instead of per-process `Hasher` values, so they compare across launches.

`finalizeInterruptedRuns` notes the runs it interrupts, and
`recoverOrphanedRunningRuns` — only it, since a user or queue cancel
leaves the live worker to compare — replays each baseline off the main
actor, appends what changed, logs `recovered=true`, and removes it.
Changes made between the crash and the relaunch are attributed to the
interrupted run. An unreadable baseline is skipped and removed; writing
a baseline removes any other the task left behind. Recovery is only
scheduled for runs that left a baseline.

+++ Tests/TaskFolderRunSnapshotTests.swift
@@ -168,7 +168,7 @@ struct TaskFolderRunSnapshotTests {
         #expect(unbudgeted?.entries.values.allSatisfy { $0.contentFingerprint == nil } == true)
 
         let stamp = Date(timeIntervalSince1970: 1_790_000_000)
-        func entry(_ fingerprint: Int?) -> TaskFolderRunSnapshot.Entry {
+        func entry(_ fingerprint: UInt64?) -> TaskFolderRunSnapshot.Entry {
             .init(size: 4, modifiedAt: stamp, statusChangedAt: stamp, fileIdentifier: 7, contentFingerprint: fingerprint)
         }
         #expect(entry(1).differs(from: entry(2)))
